+++ src/core/production_map/types/definition.rs
@@ -14,7 +14,7 @@ pub struct ProductionMapDefinition {
     #[serde(default, skip_serializing_if = "String::is_empty")]
     pub customer_name: String,
     #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub roll_count: Option<f64>,
+    pub roll_count: Option<i64>,
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub width_mm: Option<f64>,
     #[serde(default, skip_serializing_if = "Option::is_none")]
